# docs/openapi.yaml
openapi: 3.0.0
info:
    title: API de Partidos de Deportes en Equipo
    description: API para la gestión de usuarios, agrupaciones y partidos con sistema dinámico de méritos.
    version: 1.5.0
security:
    - bearerAuth: []
paths:
...
    /api/groups/{id}/matches:
        get:
            summary: Listar partidos de un grupo
            description: Obtiene todos los partidos organizados por el grupo. Solo miembros.
            operationId: listGroupMatches
            tags: [Matches]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Match" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches:
        post:
            summary: Crear un nuevo partido
            description: Organiza un encuentro deportivo. Solo Owners.
            operationId: createMatch
            tags: [Matches]
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/MatchInput" }
            responses:
                "201":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Match" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches/{id}:
        patch:
            summary: Actualizar metadatos del partido
            description: Modifica ubicación, capacidad o colores. No permitido si está bloqueado. Solo Owners.
            operationId: updateMatch
            tags: [Matches]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            requestBody:
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/MatchUpdate" }
            responses:
                "200":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Match" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches/{id}/status:
        patch:
            summary: Cambiar estado del partido
            description: Gestiona el ciclo de vida (planning -> playing -> finished). Sella reservas al pasar a 'playing'. Solo Owners.
            operationId: updateMatchStatus
            tags: [Matches]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                status: { $ref: "#/components/schemas/MatchStatus" }
            responses:
                "200":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Match" }
                "400": { $ref: "#/components/responses/BadRequest" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches/{id}/lock:
        patch:
            summary: Bloquear partido
            description: Hace el partido inmutable. Solo permitido para partidos en estado 'finished'. Solo Owners.
            operationId: lockMatch
            tags: [Matches]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "200": { description: Partido bloqueado exitosamente }
                "400": { $ref: "#/components/responses/BadRequest" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches/{id}/participants:
        get:
            summary: Listar participantes (Orden de Mérito)
            description: Obtiene la lista de inscritos ordenada dinámicamente por sus puntos de mérito históricos. Incluye cálculo de reserva.
            operationId: listParticipants
            tags: [Participants]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/MatchRegistrationWithMerit" }
        post:
            summary: Apuntarse a un partido
            description: Registra al usuario autenticado. Solo en estado 'planning'.
            operationId: joinMatch
            tags: [Participants]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "201":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/MatchRegistration" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches/{id}/participants/me:
        delete:
            summary: Salirse de un partido
            description: El usuario abandona el partido. Si es tarde (según config del grupo), se mantiene el registro con penalización.
            operationId: leaveMatch
            tags: [Participants]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "200": { description: Salida procesada }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/matches/{id}/participants/{userId}/evaluation:
        patch:
            summary: Evaluar participación
            description: Confirma si el usuario asistió y su actitud. Solo en estado 'finished'. Solo Owners.
            operationId: evaluateParticipant
            tags: [Participants]
            parameters:
                - { $ref: "#/components/parameters/Id" }
                - name: userId
                  in: path
                  required: true
                  schema: { type: string, format: uuid }
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                didPlay: { type: boolean }
                                attitude: { $ref: "#/components/schemas/PlayerAttitude" }
            responses:
                "200": { description: Evaluación guardada }
                "403": { $ref: "#/components/responses/Forbidden" }
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
        Id:
            name: id
            in: path
            required: true
            schema: { type: string, format: uuid }
    schemas:
...
        RGB:
            type: object
            properties:
                r: { type: number, minimum: 0, maximum: 1 }
                g: { type: number, minimum: 0, maximum: 1 }
                b: { type: number, minimum: 0, maximum: 1 }
        SportType: { type: string, enum: [football, basketball, padel] }
        MatchStatus: { type: string, enum: [planning, playing, finished, cancelled] }
        PlayerAttitude: { type: string, enum: [positive, negative, neutral] }
        Match:
            type: object
            properties:
                id: { type: string, format: uuid }
                groupId: { type: string, format: uuid }
                sport: { $ref: "#/components/schemas/SportType" }
                scheduledAt: { type: string, format: date-time }
                durationMinutes: { type: integer }
                capacity: { type: integer }
                location: { type: string }
                status: { $ref: "#/components/schemas/MatchStatus" }
                isLocked: { type: boolean }
                teamAColor: { $ref: "#/components/schemas/RGB" }
                teamBColor: { $ref: "#/components/schemas/RGB" }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
        MatchInput:
            type: object
            required: [groupId, sport, scheduledAt, durationMinutes, capacity, location, teamAColor, teamBColor]
            properties:
                groupId: { type: string, format: uuid }
                sport: { $ref: "#/components/schemas/SportType" }
                scheduledAt: { type: string, format: date-time }
                durationMinutes: { type: integer }
                capacity: { type: integer }
                location: { type: string }
                teamAColor: { $ref: "#/components/schemas/RGB" }
                teamBColor: { $ref: "#/components/schemas/RGB" }
        MatchUpdate:
            type: object
            properties:
                sport: { $ref: "#/components/schemas/SportType" }
                scheduledAt: { type: string, format: date-time }
                durationMinutes: { type: integer }
                capacity: { type: integer }
                location: { type: string }
                teamAColor: { $ref: "#/components/schemas/RGB" }
                teamBColor: { $ref: "#/components/schemas/RGB" }
        MatchRegistration:
            type: object
            properties:
                matchId: { type: string, format: uuid }
                userId: { type: string, format: uuid }
                registeredAt: { type: string, format: date-time }
                isReserve: { type: boolean }
                didPlay: { type: boolean, nullable: true }
                attitude: { $ref: "#/components/schemas/PlayerAttitude" }
                isLateCancellation: { type: boolean }
        MatchRegistrationWithMerit:
            allOf:
                - $ref: "#/components/schemas/MatchRegistration"
                - type: object
                  properties:
                      totalMeritPoints: { type: integer, description: Puntos acumulados en los últimos partidos }
                      isReserveCalculated: { type: boolean, description: Si el usuario es reserva según el ranking actual y capacidad }
    responses:
        BadRequest:
            description: Error en los datos.
            content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
        Forbidden:
            description: Permisos insuficientes o recurso bloqueado.
            content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...